requirements.txt
Pillow 
rembg 
opencv-python 
PyMatting
pydub
moviepy